requirements.txt
argh==0.26.2
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap4==4.0.2
Flask-Moment==0.7.0
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pathtools==0.1.2
piexif==1.1.2
pillow>=6.2.2
httplib2>=0.18.0
psycopg2==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-dotenv==0.10.3
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
visitor==0.1.3
watchdog==0.9.0
Werkzeug==0.15.4
WTForms==2.2.1
